Fix maps for reporting (#15272)
This fix ensures the render-complete event does not fire before rendering is actually completed. 

- This ensures that the ChoroplethLayer is fully loaded, by explicitly waiting until all geojson files are loaded.
- It also removes the special init() code-path for coordinate-maps.This was the cause of coordinate-maps not showing on reports.
