--- src/core_plugins/region_map/public/region_map_controller.html
@@ -1,3 +1,0 @@
-<div ng-controller="KbnRegionMapController" class="region-map-vis">
-
-</div>
