+++ src/core_plugins/region_map/public/region_map_vis.js
@@ -1,7 +1,5 @@
 import './region_map.less';
-import './region_map_controller';
 import './region_map_vis_params';
-import regionTemplate from './region_map_controller.html';
 import image from './images/icon-vector-map.svg';
 import { VisFactoryProvider } from 'ui/vis/vis_factory';
 import { CATEGORY } from 'ui/vis/vis_category';
@@ -9,16 +7,18 @@ import { VisSchemasProvider } from 'ui/vis/editors/default/schemas';
 import { VisTypesRegistryProvider } from 'ui/registry/vis_types';
 import { truncatedColorMaps } from 'ui/vislib/components/color/truncated_colormaps';
 import { mapToLayerWithId } from './util';
+import { RegionMapsVisualizationProvider } from './region_map_visualization';
 
 VisTypesRegistryProvider.register(function RegionMapProvider(Private, regionmapsConfig, config) {
   const VisFactory = Private(VisFactoryProvider);
   const Schemas = Private(VisSchemasProvider);
+  const RegionMapsVisualization = Private(RegionMapsVisualizationProvider);
 
   const vectorLayers = regionmapsConfig.layers.map(mapToLayerWithId.bind(null, 'self_hosted'));
   const selectedLayer = vectorLayers[0];
   const selectedJoinField = selectedLayer ? vectorLayers[0].fields[0] : null;
 
-  return VisFactory.createAngularVisualization({
+  return VisFactory.createBaseVisualization({
     name: 'region_map',
     title: 'Region Map',
     implementsRenderComplete: true,
@@ -34,10 +34,12 @@ VisTypesRegistryProvider.register(function RegionMapProvider(Private, regionmaps
         selectedLayer: selectedLayer,
         selectedJoinField: selectedJoinField,
         isDisplayWarning: true,
-        wms: config.get('visualization:tileMap:WMSdefaults')
-      },
-      template: regionTemplate,
+        wms: config.get('visualization:tileMap:WMSdefaults'),
+        mapZoom: 2,
+        mapCenter: [0, 0],
+      }
     },
+    visualization: RegionMapsVisualization,
     editorConfig: {
       optionsTemplate: '<region_map-vis-params></region_map-vis-params>',
       collections: {
